+++ packages/minimongo/selector.js
@@ -653,7 +653,7 @@ var ELEMENT_OPERATORS = {
         throw Error("$elemMatch need an object");
 
       var subMatcher, isDocMatcher;
-      if (isOperatorObject(operand)) {
+      if (isOperatorObject(operand, true)) {
         subMatcher = compileValueSelector(operand, matcher);
         isDocMatcher = false;
       } else {